Tribute Page
Check out the Tribute  Page here:
👩‍💼 Key Features:
✅ Responsive Design: Ensuring that the Tribute Page is accessible across all devices.
✅ Thoughtful Layout: Balancing text and visuals to engage and inform visitors.
✅ CSS Styling: Employing CSS to enhance aesthetics and user experience.
✅ Dedication: Paying homage to an individual who has inspired many.
Building this Tribute Page was a heartfelt endeavor that allowed me to not only refine my web development skills but also to connect with a remarkable story. It serves as a reminder of the power of technology to unite us in commemorating those who've left their mark on the world.
I encourage you to visit the page, take a moment to read about our inspiring subject, and reflect on the impact they've had. Feel free to share your thoughts and feedback as well – I'd love to hear from you!